1 .. 10

+++ pointers_arrays_strings/100-atoi.c
@@ -8,19 +8,19 @@
  */
 int _atoi(char *s)
 {
-int sign = 1;
-int result = 0;
-int i = 0;
+	int sign = 1;
+	long int result = 0;
+	int i = 0;
 
 	while (s[i] != '\0')
 	{
-	if (s[i] == '-')
-	sign *= -1;
-	else if (s[i] >= '0' && s[i] <= '9')
-	result = (result * 10) + (s[i] - '0');
-	else if (result != 0)
-	break;
-	i++;	
+		if (s[i] == '-')
+			sign *= -1;
+		else if (s[i] >= '0' && s[i] <= '9')
+			result = (result * 10) + (s[i] - '0');
+		else if (result != 0)
+			break;
+		i++;
 	}
 
 	return (result * sign);
